Extract page routing & per-page initialization into dedicated route modules
